* commit '30772577022aeda9a9db7bcc00317445d4a5a54a':
  无jira任务 查询filter解析过程修复 * 目前仅appId 和 yearMonth 组合支持快速查询 * 其他情况全局查询,后续优化
  无jira任务，合并代码
  无jira任务，解决冲突
  KERNEL-3757 单机版本调试测试 1、bucket调整为内存获取 2、接入内存bucket 3、删除块时同时移除内存bucket，数据库依靠触发器
  KERNEL-3757 单机版本调试测试 1、bucket调整为内存获取 2、接入内存bucket 3、删除块时同时移除内存bucket，数据库依靠触发器
  KERNEL-3757 单机版本调试测试 1、collate逻辑调整 2、bugfix多任务并行移植
  KERNEL-3757 单机版本调试测试 1、bucket调整为内存获取 2、接入内存bucket 3、删除块时同时移除内存bucket，数据库依靠触发器
  无jira任务 修改collate bug
  无jira任务 修改collate bug
  无jira任务 collate 适配line方式以及不需要collate的不新建块
  KERNEL-3632 jdbc模块兼容适配，并尝试同步移植到日志引擎中 1、jdbc模块兼容适配，修改报错
  REPORT-28702【rds分析需求】03_200318_dashboard操作统计 顺便修复delete报错 archive让SSL不疯狂刷绿色的SSL warning
  KERNEL-3452 无用代码移除、复杂代码整理重写 1、修改注释
  KERNEL-3452 无用代码移除、复杂代码整理重写 1、serverservice接入 2、swiftservice接入 3、container调整测试 4、delete接入ProcessHandler，同时删除后若allshow为空，直接删除文件和配置
  KERNEL-3452 无用代码移除、复杂代码整理重写 1、segment相关service和container重写 2、接入import、collate
  KERNEL-3452 1、无用代码移除、复杂代码整理重写 2、数据库相关重写
  REPORT-27422 阿里云秘钥泄露 1、删除oss配置
  REPORT-27422 阿里云秘钥泄露 1、删除oss配置
  KERNEL-2968 冗余、无用代码整理、移动                           1、部分无用代码删除                           2、v3版本数据增加yearmonth的hash分块
  KERNEL-2968 冗余、无用代码整理、移动                           1、部分无用代码删除                           2、v3版本数据增加yearmonth的hash分块

+++ swift-base/swift-base-io/src/test/java/com/fr/swift/cube/io/impl/BaseByteArrayWriterTest.java
@@ -60,7 +60,7 @@ public class BaseByteArrayWriterTest {
         verify(dataWriter).put(1, (byte) 2);
         verify(dataWriter).put(2, (byte) 3);
 
-        assertEquals(3L, Whitebox.getInternalState(byteArrayWriter, "curPos"));
+        assertEquals(3L, (long) Whitebox.getInternalState(byteArrayWriter, "curPos"));
 
         byteArrayWriter.put(1, null);
 
@@ -68,13 +68,13 @@ public class BaseByteArrayWriterTest {
         verify(lenWriter).put(1, 0);
         verifyNoMoreInteractions(dataWriter);
 
-        assertEquals(3L, Whitebox.getInternalState(byteArrayWriter, "curPos"));
+        assertEquals(3L, (long) Whitebox.getInternalState(byteArrayWriter, "curPos"));
 
         when(lastPosReader.isReadable()).thenReturn(true);
         when(lastPosReader.get(0)).thenReturn(3L);
 
         byteArrayWriter = new BaseByteArrayWriter(dataWriter, posWriter, lenWriter, false, lastPosWriter, lastPosReader);
-        assertEquals(3L, Whitebox.getInternalState(byteArrayWriter, "curPos"));
+        assertEquals(3L, (long) Whitebox.getInternalState(byteArrayWriter, "curPos"));
     }
 
     @Test
@@ -92,7 +92,7 @@ public class BaseByteArrayWriterTest {
         verify(dataWriter).put(1, (byte) 2);
         verify(dataWriter).put(2, (byte) 3);
 
-        assertEquals(3L, Whitebox.getInternalState(byteArrayWriter, "curPos"));
+        assertEquals(3L, (long) Whitebox.getInternalState(byteArrayWriter, "curPos"));
 
         output = byteArrayWriter.putStream(1);
         output.close();
@@ -101,13 +101,13 @@ public class BaseByteArrayWriterTest {
         verify(lenWriter).put(1, 0);
         verifyNoMoreInteractions(dataWriter);
 
-        assertEquals(3L, Whitebox.getInternalState(byteArrayWriter, "curPos"));
+        assertEquals(3L, (long) Whitebox.getInternalState(byteArrayWriter, "curPos"));
 
         when(lastPosReader.isReadable()).thenReturn(true);
         when(lastPosReader.get(0)).thenReturn(3L);
 
         byteArrayWriter = new BaseByteArrayWriter(dataWriter, posWriter, lenWriter, false, lastPosWriter, lastPosReader);
-        assertEquals(3L, Whitebox.getInternalState(byteArrayWriter, "curPos"));
+        assertEquals(3L, (long) Whitebox.getInternalState(byteArrayWriter, "curPos"));
     }
 
     @Test
@@ -116,7 +116,7 @@ public class BaseByteArrayWriterTest {
         Whitebox.setInternalState(byteArrayWriter, "curPos", 1L);
 
         byteArrayWriter.resetContentPosition();
-        assertEquals(0L, Whitebox.getInternalState(byteArrayWriter, "curPos"));
+        assertEquals(0L, (long) Whitebox.getInternalState(byteArrayWriter, "curPos"));
     }
 
     @Test
